src/generator_service/requirements.txt
fastapi[all]
pydantic
kafka-python
